+++ python/cudf/bindings/reduce.pyx
@@ -44,7 +44,7 @@ def apply_reduce(reduction, col):
 
     func = bytes(reduction, encoding="UTF-8")
 
-    outsz = gdf_reduce_optimal_output_size()
+    outsz = gdf_reduction_get_intermediate_output_size()
     out = rmm.device_array(outsz, dtype=col.dtype)
     cdef uintptr_t out_ptr = get_ctype_ptr(out)